pytorch_lr_classifier example now functional
